Layout changes and few validations added 02.28

+++ src/main/java/pl/orionproject/controller/AdminController.java
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import pl.orionproject.DataTransferObjects.CategoryDto;
 import pl.orionproject.DataTransferObjects.ItemDto;
+import pl.orionproject.component.CategoryValidator;
+import pl.orionproject.component.ItemValidator;
 import pl.orionproject.model.Category;
 import pl.orionproject.model.Item;
 import pl.orionproject.repository.CategoryRepository;
@@ -29,6 +31,12 @@ public class AdminController {
     @Autowired
     ItemRepository itemRepository;
 
+    @Autowired
+    ItemValidator itemValidator;
+
+    @Autowired
+    CategoryValidator categoryValidator;
+
     @GetMapping("/admin")
     public String viewAdminPage() {
         return "admin";
@@ -46,15 +54,18 @@ public class AdminController {
 
     @GetMapping("/admin/additem")
     public String viewAddItemPage(Model model) {
-//        model.addAttribute("categories", categoryRepository.findAll());
         model.addAttribute("allitems", itemRepository.findAll());
         return "additem";
     }
 
     @PostMapping("/admin/additem")
     public String addItem(@ModelAttribute("item") ItemDto itemDto) {
-        itemService.addItem(itemDto);
-        return "redirect:/";
+        if (itemValidator.isItemExists(itemDto.getItemName())) {
+            return "redirect:/admin/additem?fail";
+        } else {
+            itemService.addItem(itemDto);
+            return "redirect:/";
+        }
     }
 
     @GetMapping("/admin/item/delete/{id}")
@@ -71,25 +82,31 @@ public class AdminController {
 
     @PostMapping("/admin/addcategories")
     public String addCategories(@ModelAttribute("category") CategoryDto categoryDto) {
+        if (categoryValidator.isCategoryExists(categoryDto.getCategoryName())) {
+            return "redirect:/admin/addcategories?fail";
+        }
         categoryRepository.save(new Category(categoryDto.getCategoryName()));
         return "redirect:/admin/addcategories";
     }
 
     @GetMapping("/admin/category/delete/{id}")
     public String deleteCategory(@PathVariable Long id) {
+        if (categoryValidator.isCategoryExists(id)) {
+            return "redirect:/admin/addcategories?fail";
+        }
         categoryRepository.deleteCategoriesByCategoryId(id);
         return "redirect:/admin/addcategories";
     }
 
     @GetMapping("/admin/category/modify/{id}")
-    public String modifyCategory (@PathVariable Long id, Model model) {
+    public String modifyCategory(@PathVariable Long id, Model model) {
         Category category = categoryRepository.findByCategoryId(id);
         model.addAttribute("categorytomodify", category);
         return "categorymodify";
     }
 
-    @PostMapping("/admin/category/modify/{id}")
-    public String modifyCategory (@PathVariable Long id, @ModelAttribute("category") CategoryDto categoryDto) {
+    @PostMapping("/admin/category/modify/{id}") //
+    public String modifyCategory(@PathVariable Long id, @ModelAttribute("category") CategoryDto categoryDto) {
         Category category = categoryRepository.findByCategoryId(id);
         category.setCategoryName(categoryDto.getCategoryName());
         categoryRepository.save(category);
@@ -97,14 +114,14 @@ public class AdminController {
     }
 
     @GetMapping("/admin/item/modify/{id}")
-    public String modifyItem (@PathVariable Long id, Model model) {
+    public String modifyItem(@PathVariable Long id, Model model) {
         Item item = itemRepository.findItemById(id);
         model.addAttribute("itemtomodify", item);
         return "itemmodify";
     }
 
-    @PostMapping("/admin/item/modify/{id}")
-    public String modifyItem (@PathVariable Long id, @ModelAttribute("item") ItemDto itemDto) {
+    @PostMapping("/admin/item/modify/{id}") //
+    public String modifyItem(@PathVariable Long id, @ModelAttribute("item") ItemDto itemDto) {
         Item item = itemRepository.findItemById(id);
         item.setItemName(itemDto.getItemName());
         item.setPrice(itemDto.getPrice());
